requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
